+++ ProjectBarnabus/src/GameEngine/Game.h
@@ -5,7 +5,7 @@ public:
 	Game();
 	~Game();
 	virtual bool LoadGameContent();
-	virtual bool Update(double deltaTime);
-	virtual bool Render(double deltaTime);
+	virtual bool Update(float deltaTime);
+	virtual bool Render(float deltaTime);
 
 };
